# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aerolit документация АПИ",
        "description": "Документация",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/items/": {
            "post": {
                "tags": [
                    "items"
                ],
                "summary": "Вытащить все товары",
                "description": "Вытащить все товары",
                "operationId": "viewitem",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name,  category, popular"
                                ],
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    },
                                    "popular": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Урна"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CallBack"
                    }
                }
            }
        },
        "/api/items/single": {
            "post": {
                "tags": [
                    "items"
                ],
                "summary": "Вытащить отдельный товар",
                "description": "Вытащить отдельный товар",
                "operationId": "singleitem",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CallBack"
                    }
                }
            }
        },
        "/api/review/": {
            "get": {
                "tags": [
                    "review"
                ],
                "summary": "Отзывы показы",
                "operationId": "reviewindex",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "tags": [
                    "review"
                ],
                "summary": "Заполнить отзыв",
                "description": "Заполнить отзыв",
                "operationId": "createreview",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "api_token, start, content"
                                ],
                                "properties": {
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Zx9HfKnJ1DZETiScaA0nU8Xm5E2yTPJGHDAXSsfdgfW8ioc6Pxzof3FotPQQ"
                                    },
                                    "start": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    },
                                    "content": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CallBack"
                    }
                }
            }
        }
    }
}